requirements.txt
asgiref==3.7.0
beautifulsoup4==4.12.2
Django==4.2.1
django-bootstrap-v5==1.0.11
django-extensions==3.2.1
flake8==6.0.0
isort==5.12.0
mccabe==0.7.0
pycodestyle==2.10.0
pycryptodome==3.18.0
pyflakes==3.0.1
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.0
